requirements.txt
rich
arrow
bilibili-api-python